fix(julia): preserve fractional stoichiometry in Catalyst→ESM reverse conversion (gt-3ai5)
The `EarthSciSerialization.ReactionSystem(::Catalyst.ReactionSystem)` method
used `Int()` to coerce substoich/prodstoich, which raised `InexactError` for
fractional stoichiometry common in atmospheric chemistry (e.g.
`CH3O2+CH3O2 -> 2.0 CH2O + 0.8 HO2`). The schema and
`StoichiometryEntry.stoichiometry::Float64` already support Float64; only the
reverse method's local `Dict{String,Int}` containers were stale. Switch to
`Dict{String,Float64}` and `Float64(...)` so fractional coefficients survive
the round-trip intact.

+++ packages/EarthSciSerialization.jl/ext/EarthSciSerializationCatalystExt.jl
@@ -213,19 +213,19 @@ function EarthSciSerialization.ReactionSystem(rs::Catalyst.ReactionSystem)
 
     reactions = Reaction[]
     for rxn in Catalyst.reactions(rs)
-        reactants = Dict{String,Int}()
+        reactants = Dict{String,Float64}()
         if !isempty(rxn.substrates)
             for (i, s) in enumerate(rxn.substrates)
                 name = _strip_time(string(Catalyst.getname(s)))
-                stoich = length(rxn.substoich) >= i ? Int(rxn.substoich[i]) : 1
+                stoich = length(rxn.substoich) >= i ? Float64(rxn.substoich[i]) : 1.0
                 reactants[name] = stoich
             end
         end
-        products = Dict{String,Int}()
+        products = Dict{String,Float64}()
         if !isempty(rxn.products)
             for (i, pr) in enumerate(rxn.products)
                 name = _strip_time(string(Catalyst.getname(pr)))
-                stoich = length(rxn.prodstoich) >= i ? Int(rxn.prodstoich[i]) : 1
+                stoich = length(rxn.prodstoich) >= i ? Float64(rxn.prodstoich[i]) : 1.0
                 products[name] = stoich
             end
         end
